Limpando o codigo de warning

+++ 05_FirstStepsInJavaWithSpringBoot/rest-with-spring-boot-and-java/src/main/java/br/com/yuri/services/math/SimpleMathService.java
@@ -23,6 +23,6 @@ public class SimpleMathService {
         }
 
         public Double squareRoot(Double number) {
-            return (Double) Math.sqrt(number);
+            return Math.sqrt(number);
         }
 }
